profile_cookies/requirements.txt
pycryptodome>=3.11.0